+++ src/test/java/io/github/astrapi69/swing/JMCheckBoxTest.java
@@ -22,7 +22,7 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package io.github.astrapi69.swing.check.model;
+package io.github.astrapi69.swing;
 
 import java.awt.*;
 
@@ -33,9 +33,10 @@ import io.github.astrapi69.model.api.Model;
 import io.github.astrapi69.model.api.SerializableConsumer;
 import io.github.astrapi69.model.api.SerializableSupplier;
 import io.github.astrapi69.swing.JMCheckBox;
+import io.github.astrapi69.swing.check.model.CheckedModelBean;
 import io.github.astrapi69.window.adapter.CloseWindow;
 
-public class JCheckBoxDecoratorTest
+public class JMCheckBoxTest
 {
 	public static void main(String[] args)
 	{
@@ -45,22 +46,22 @@ public class JCheckBoxDecoratorTest
 		checkedModelBean = CheckedModelBean.builder().build();
 		checkBox = new JMCheckBox("Check me");
 
-		SerializableSupplier<Boolean> getter = () -> {
-			Boolean checked = checkBox.isSelected();
-			checkedModelBean.setChecked(checked);
-			return checked;
-		};
-		SerializableConsumer<Boolean> setter = (checked) -> {
-			checkBox.setSelected(checked);
-			checkedModelBean.setChecked(checked);
-		};
+//		SerializableSupplier<Boolean> getter = () -> {
+//			Boolean checked = checkBox.isSelected();
+//			checkedModelBean.setChecked(checked);
+//			return checked;
+//		};
+//		SerializableConsumer<Boolean> setter = (checked) -> {
+//			checkBox.setSelected(checked);
+//			checkedModelBean.setChecked(checked);
+//		};
 		final Model<Boolean> booleanModel =
 			// PropertyModel.of(checkedModelBean, "checked");
 			// LambdaModel.of(getter, setter);
 			LambdaModel.of(checkedModelBean::isChecked, checkedModelBean::setChecked);
 		checkBox.setPropertyModel(booleanModel);
 
-		final Frame frame = new Frame("JCheckBoxDecoratorTest");
+		final Frame frame = new Frame("JMCheckBoxTest");
 		JButton buttonCheck = new JButton("check it");
 		buttonCheck.addActionListener(e -> {
 			Boolean selected = ((JMCheckBox)checkBox).getPropertyModel().getObject();