+++ src/test/java/io/github/astrapi69/swing/JTextFieldTest.java
@@ -40,7 +40,7 @@ public class JTextFieldTest
 		JTextField textField = new JTextField("fff", 20);
 
 		final PropertyModel<String> propertyModel = PropertyModel.<String> of(textField, "text");
-		final Frame frame = new Frame("JCheckBoxDecoratorTest");
+		final Frame frame = new Frame("JMCheckBoxTest");
 		JButton buttonCheck = new JButton("check it");
 		buttonCheck.addActionListener(e -> {
 			String selected = propertyModel.getObject();