+++ src/test/java/io/github/astrapi69/swing/base/ApplicationPanelFrameExample.java
@@ -29,6 +29,7 @@ import java.awt.*;
 import javax.swing.*;
 import javax.swing.border.Border;
 
+import io.github.astrapi69.swing.plaf.LookAndFeels;
 import org.jdesktop.swingx.MultiSplitLayout;
 
 import io.github.astrapi69.swing.button.IconButtonFactory;
@@ -54,6 +55,7 @@ public class ApplicationPanelFrameExample
 	public ApplicationPanelFrameExample(String title)
 	{
 		super(title);
+		setDefaultLookAndFeel(LookAndFeels.NIMBUS, this);
 	}
 
 	/**